# app/static/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "InfraMusicStore API",
    "version": "1.0.0",
    "description": "API REST pour gérer artistes, albums, genres et pistes."
  },
  "paths": {
    "/api/artists": {
      "get": {
        "summary": "Liste des artistes",
        "responses": {
          "200": {
            "description": "Liste retournée"
          }
        }
      },
      "post": {
        "summary": "Créer un artiste",
        "requestBody": {
          "required": true
        },
        "responses": {
          "201": {
            "description": "Artiste créé"
          }
        }
      }
    },
    "/api/artists/{artistId}": {
      "get": {
        "summary": "Récupérer un artiste",
        "parameters": [
          {
            "name": "artistId",
            "in": "path",
            "required": true,
            "schema": { "type": "integer" }
          }
        ],
        "responses": {
          "200": { "description": "Artiste trouvé" },
          "404": { "description": "Non trouvé" }
        }
      },
      "put": {
        "summary": "Mettre à jour un artiste",
        "parameters": [
          {
            "name": "artistId",
            "in": "path",
            "required": true,
            "schema": { "type": "integer" }
          }
        ],
        "responses": {
          "200": { "description": "Artiste mis à jour" }
        }
      },
      "delete": {
        "summary": "Supprimer un artiste",
        "parameters": [
          {
            "name": "artistId",
            "in": "path",
            "required": true,
            "schema": { "type": "integer" }
          }
        ],
        "responses": {
          "200": { "description": "Artiste supprimé" },
          "404": { "description": "Non trouvé" }
        }
      }
    },

    "/api/albums": {
      "get": { "summary": "Liste des albums", "responses": { "200": { "description": "OK" } } },
      "post": { "summary": "Créer un album", "responses": { "201": { "description": "Créé" } } }
    },
    "/api/albums/{albumId}": {
      "get": { "summary": "Récupérer un album", "responses": { "200": { "description": "OK" }, "404": { "description": "Non trouvé" } } },
      "put": { "summary": "Mettre à jour un album", "responses": { "200": { "description": "OK" } } },
      "delete": { "summary": "Supprimer un album", "responses": { "200": { "description": "OK" }, "404": { "description": "Non trouvé" } } }
    },

    "/api/genres": {
      "get": { "summary": "Liste des genres", "responses": { "200": { "description": "OK" } } },
      "post": { "summary": "Créer un genre", "responses": { "201": { "description": "Créé" } } }
    },
    "/api/genres/{genreId}": {
      "get": { "summary": "Récupérer un genre", "responses": { "200": { "description": "OK" }, "404": { "description": "Non trouvé" } } },
      "put": { "summary": "Mettre à jour un genre", "responses": { "200": { "description": "OK" } } },
      "delete": { "summary": "Supprimer un genre", "responses": { "200": { "description": "OK" }, "404": { "description": "Non trouvé" } } }
    },

    "/api/tracks": {
      "get": { "summary": "Liste des pistes", "responses": { "200": { "description": "OK" } } },
      "post": { "summary": "Créer une piste", "responses": { "201": { "description": "Créé" } } }
    },
    "/api/tracks/{trackId}": {
      "get": { "summary": "Récupérer une piste", "responses": { "200": { "description": "OK" }, "404": { "description": "Non trouvé" } } },
      "put": { "summary": "Mettre à jour une piste", "responses": { "200": { "description": "OK" } } },
      "delete": { "summary": "Supprimer une piste", "responses": { "200": { "description": "OK" }, "404": { "description": "Non trouvé" } } }
    }
  }
}